Finally got state #4 to work, changed some background assets, other minor changes
In state 4, player is able to pick up food from the counter and deliver it to the customer's table. Once doing that, the speech bubble showing what the customer's order was disappears and the instance of the food also disappears to signify customer is eating it. Still needs further changes but entire order sequence is now functional.